feat(mobile): ParticipationBar at cluster/ path
Moves the affected/observing stat row from src/components/clusters/
(plural) to src/components/cluster/ (singular) per the mobile-01d
session task list. Pure presentation component — no behaviour changes,
no business logic.

Adds:
  - accessibilityRole="summary"
  - accessibilityLabel that reads both counts together for screen readers
  - Explicit return type annotation (React.ReactElement)
  - StyleSheet color values normalised to uppercase hex (matches the
    rest of the post-01a screens)

Action buttons (I'm Affected / I'm Observing / etc.) are NOT moved
into this component. They live inline in clusters/[id].tsx because
they need access to:
  - session-local participation state (which the API doesn't expose)
  - the 2-minute context window timer
  - the device hash for the participate mutation
  - Crypto.digestStringAsync for idempotency keys

Splitting them out would require lifting all of that state up. The
stat row is presentation-only and benefits from being a small reusable
component; the action buttons are tightly coupled to screen state.

Removes: src/components/clusters/ParticipationBar.tsx (only consumer
was clusters/[id].tsx, which is being rewritten in the next commit).

Verified: only out-of-scope errors remain (clusters/[id].tsx import
and clusters/[id]/restore.tsx restoration values — both rewritten
in subsequent mobile-01d commits).
